requirements.txt
alembic==1.17.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
contourpy==1.3.3
cycler==0.12.1
docker==7.1.0
entrypoints==0.4
Flask==3.1.2
fonttools==4.60.1
gitdb==4.0.12
GitPython==3.1.45
graphene==3.4.3
graphql-core==3.2.7
graphql-relay==3.2.0
gunicorn==22.0.0
idna==3.11
importlib_metadata==7.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
Mako==1.3.10
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
mlflow==2.14.3
numpy==1.26.4
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==24.2
pandas==2.3.3
pillow==12.0.0
protobuf==4.25.8
pyarrow==14.0.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.3
querystring-parser==1.2.4
requests==2.32.5
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
SQLAlchemy==2.0.44
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wheel==0.45.1
xgboost==3.1.1
zipp==3.23.0
